tokenizers::models::unigram

Struct Unigram

pub struct Unigram {
    pub min_score: f64,
    /* private fields */
}

A Unigram model to encode sentences.

Fields
min_score: f64
Implementations

impl Unigram

pub fn from( vocab: Vec<(String, f64)>, unk_id: Option<usize>, byte_fallback: bool, ) -> Result<Self>

Create a Unigram model from a given vocabulary. Vocabulary are the various tokens and their associated score which is a sort of a logprob of their frequency, which will enable tokenization and sampling. unk_id, is the index within the vocabulary. For now Unigram requires at least unk because we might find a never seen char. Further versions might allow that part to be hidden.

pub fn byte_fallback(&self) -> bool

pub fn encode(&self, sentence: &str) -> Result<Vec<String>>

This functions take a String, and will encode it in a Vec of Strings, of the best tokenization available to the current model.

use tokenizers::models::unigram::Unigram;

let pieces = vec![
    ("<unk>".to_string(), 0.0),
    ("a".to_string(), 0.0),
    ("b".to_string(), 0.0),
    ("c".to_string(), 0.0),
    ("d".to_string(), 0.0),
    ("cd".to_string(), 1.0),
    ("ab".to_string(), 2.0),
    ("abc".to_string(), 5.0),
    ("abcd".to_string(), 10.0),
];
let model = Unigram::from(pieces, Some(0), false).unwrap();
let result = model.encode("abcdacdxx").unwrap();
assert_eq!(result, vec!["abcd", "a", "cd", "xx"]);

pub fn iter(&self) -> UnigramIterator<'_>

Iterate of vocabulary of the model as a pair of (token, score).

pub fn load<P: AsRef<Path>>(path: P) -> Result<Unigram>

Loads a SentencePiece output model after being trained by tokenizers. After that you can use the model with tokenizers library.

use tokenizers::models::unigram::Unigram;
use std::path::Path;

let model = Unigram::load("mymodel-unigram.json").unwrap();

pub fn clear_cache(&mut self)

Clears the internal cache

pub fn resize_cache(&mut self, capacity: usize)

Resize the cache