tokenizers::tokenizer::pattern

Trait Pattern

pub trait Pattern {
    // Required method
    fn find_matches(&self, inside: &str) -> Result<Vec<(Offsets, bool)>>;
}

Pattern used to split a NormalizedString

Required Methods

fn find_matches(&self, inside: &str) -> Result<Vec<(Offsets, bool)>>

Slice the given string in a list of pattern match positions, with a boolean indicating whether this is a match or not.

This method must cover the whole string in its outputs, with contiguous ordered slices.