tokenizers::tokenizer

Struct AddedToken

pub struct AddedToken {
    pub content: String,
    pub single_word: bool,
    pub lstrip: bool,
    pub rstrip: bool,
    pub normalized: bool,
    pub special: bool,
}

Represent a token added by the user on top of the existing Model vocabulary. AddedToken can be configured to specify the behavior they should have in various situations like:

Whether they should only match single words

Whether to include any whitespace on its left or right

Fields
content: String

The content of the added token

single_word: bool

Whether this token must be a single word or can break words

lstrip: bool

Whether this token should strip whitespaces on its left

rstrip: bool

Whether this token should strip whitespaces on its right

normalized: bool

Whether this token should be normalized

special: bool

Whether this token is special

Implementations

impl AddedToken

pub fn from<S: Into<String>>(content: S, special: bool) -> Self

Build this token from the given content, specifying if it is intented to be a special token. Special tokens are not normalized by default.

pub fn single_word(self, single_word: bool) -> Self

Specify whether this token should only match on whole single words, and never part of a word.

pub fn lstrip(self, lstrip: bool) -> Self

Specify whether this token should include all the whitespaces on its left, in order to strip them out.

pub fn rstrip(self, rstrip: bool) -> Self

Specify whether this token should include all the whitespaces on its right, in order to strip them out.

pub fn normalized(self, normalized: bool) -> Self

Specify whether this token should be normalized and match against its normalized version in the input text.

pub fn special(self, special: bool) -> Self

Specify whether this token is special, meaning if it should be skipped when decoding