tokenizers::tokenizer

Trait PreTokenizer

pub trait PreTokenizer {
    // Required method
    fn pre_tokenize(&self, pretokenized: &mut PreTokenizedString) -> Result<()>;
}

The PreTokenizer is in charge of doing the pre-segmentation step. It splits the given string in multiple substrings, keeping track of the offsets of said substrings from the NormalizedString. In some occasions, the PreTokenizer might need to modify the given NormalizedString to ensure we can entirely keep track of the offsets and the mapping with the original string.

Required Methods

fn pre_tokenize(&self, pretokenized: &mut PreTokenizedString) -> Result<()>