Trait tokenizers::tokenizer::Tokenizer

pub trait Tokenizer<T: Vocab> {
    fn vocab(&self) -> &T;
    fn tokenize_to_tokens(&self, text: TokenRef<'_>) -> Vec<Token>;

    fn tokenize<S: AsRef<str>>(&self, text: S) -> Vec<String> { ... }
    fn tokenize_with_offsets<S: AsRef<str>>(&self, text: S) -> TokensWithOffsets { ... }
    fn tokenize_list<S, ST>(&self, text_list: S) -> Vec<Vec<String>>
    where
        S: AsRef<[ST]>,
        ST: AsRef<str>,
    { ... }
    fn tokenize_list_with_offsets<S, ST>(
        &self, 
        text_list: S
    ) -> Vec<TokensWithOffsets>
    where
        S: AsRef<[ST]>,
        ST: AsRef<str>,
    { ... }
    fn convert_tokens_to_ids<S, ST>(&self, tokens: S) -> Vec<i64>
    where
        S: AsRef<[ST]>,
        ST: AsRef<str>,
    { ... }
    fn encode<S: AsRef<str>>(
        &self, 
        text_1: S, 
        text_2: Option<S>, 
        max_len: usize, 
        truncation_strategy: &TruncationStrategy, 
        stride: usize
    ) -> TokenizedInput { ... }
    fn encode_list<S, ST>(
        &self, 
        text_list: S, 
        max_len: usize, 
        truncation_strategy: &TruncationStrategy, 
        stride: usize
    ) -> Vec<TokenizedInput>
    where
        S: AsRef<[ST]>,
        ST: AsRef<str>,
    { ... }
    fn encode_pair_list<S, ST>(
        &self, 
        text_list: S, 
        max_len: usize, 
        truncation_strategy: &TruncationStrategy, 
        stride: usize
    ) -> Vec<TokenizedInput>
    where
        S: AsRef<[(ST, ST)]>,
        ST: AsRef<str>,
    { ... }
    fn clean_up_tokenization(&self, input_string: String) -> String { ... }
    fn build_input_with_special_tokens(
        &self, 
        tokens_ids_with_offsets_1: TokenIdsWithOffsets, 
        tokens_ids_with_offsets_2: Option<TokenIdsWithOffsets>
    ) -> TokenIdsWithSpecialTokens { ... }
}

Base trait for tokenizers

Required Methods

fn vocab(&self) -> &T

returns a reference to the tokenizer vocabulary

fn tokenize_to_tokens(&self, text: TokenRef<'_>) -> Vec<Token>

Tokenize a TokenRef, returning a sequence of tokens

Parameters

text (TokenRef): TokenRef to tokenize (this is especially useful for nested tokenization, where a tokenizer is called on the ouput of a pre-tokenizer, such as BERT).

Returns

Vec<Token> tokenization of the original TokenRef

Example

Provided Methods

fn tokenize<S: AsRef<str>>(&self, text: S) -> Vec<String>

Tokenize a string, returns a vector of tokens as strings. Use tokenize_with_offsets or tokenize_to_tokens to return offset information.

Parameters

text : text (string-like) to tokenize

Returns

Vec<String> containing the tokens string representation

Example

fn tokenize_with_offsets<S: AsRef<str>>(&self, text: S) -> TokensWithOffsets

Tokenize a string, returning tokens with offset information

Parameters

text : text (string-like) to tokenize

Returns

TokensWithOffsets with the tokens and their offset information

Example

fn tokenize_list<S, ST>(&self, text_list: S) -> Vec<Vec<String>> where
S: AsRef<[ST]>,
ST: AsRef<str>,

Tokenize a list of strings, returning tokens with offset information

Parameters

text_list: list of strings to tokenize

Returns

Vec<Vec<String>> with the token strings representation

Example

fn tokenize_list_with_offsets<S, ST>(
&self,
text_list: S
) -> Vec<TokensWithOffsets> where
S: AsRef<[ST]>,
ST: AsRef<str>,

Tokenize a list of strings, where each corresponds to for example a sentence, returns a vector of TokensWithOffsets containing the tokens and their offset information. This calls tokenize_with_offsets on the list provided.

Parameters

text_list: list of strings to tokenize

Returns

Vec<TokensWithOffsets> with the token strings representation and offsets

Example

fn convert_tokens_to_ids<S, ST>(&self, tokens: S) -> Vec<i64> where
S: AsRef<[ST]>,
ST: AsRef<str>,

Convert a slice of string-like to a vector ot token indices

Parameters

tokens: list of token string-like to convert to ids

Returns

Vec<i64> with the token indices

Example

fn encode<S: AsRef<str>>(
&self,
text_1: S,
text_2: Option<S>,
max_len: usize,
truncation_strategy: &TruncationStrategy,
stride: usize
) -> TokenizedInput

Encode a string-like (tokenization followed by encoding)

Parameters

text_1: input text (string-like) to encode

text_2: optional additional input text (string-like) to encode. When provided, both texts are combined into a single encoding by using the build_input_with_special_tokens method.

max_len (usize): maximum combined sequence length. If the combined encoding would exceed this max_len, the encoding is truncated following the TruncationStrategy provided.

truncation_strategy (&TruncationStrategy): strategy to follow for the truncation, if required

stride (usize): amount of tokens to shift the input by if truncation is required (allowing for the generation of overlapping sequences with overflowing tokens)

Returns

TokenizedInput containing the encoding output (token indices, token types, segment ids, ovrflowing tokens and special token mask)

Example

fn encode_list<S, ST>(
&self,
text_list: S,
max_len: usize,
truncation_strategy: &TruncationStrategy,
stride: usize
) -> Vec<TokenizedInput> where
S: AsRef<[ST]>,
ST: AsRef<str>,

Encode a sequence of string-like texts (tokenization followed by encoding). Not that in contrast with encode optional second text, each text provided is encoded independently.

Parameters

text_list: sequence of input text (&str) to encode combined into a single encoding by using the build_input_with_special_tokens method.

max_len (usize): maximum combined sequence length. If the combined encoding would exceed this max_len, the encoding is truncated following the TruncationStrategy provided.

truncation_strategy (&TruncationStrategy): strategy to follow for the truncation, if required

stride (usize): amount of tokens to shift the input by if truncation is required (allowing for the generation of overlapping sequences with overflowing tokens)

Returns

Vec<TokenizedInput> containing the encoding output (token indices, token types, segment ids, ovrflowing tokens and special token mask) for each provided text

Example

fn encode_pair_list<S, ST>(
&self,
text_list: S,
max_len: usize,
truncation_strategy: &TruncationStrategy,
stride: usize
) -> Vec<TokenizedInput> where
S: AsRef<[(ST, ST)]>,
ST: AsRef<str>,

Encode a sequence of string-like text pairs (tokenization followed by encoding). This combines with encode with the list processing of encode_list.

Parameters

text_list: sequence of input text (&str) to encode combined into a single encoding by using the build_input_with_special_tokens method.

max_len (usize): maximum combined sequence length. If the combined encoding would exceed this max_len, the encoding is truncated following the TruncationStrategy provided.

truncation_strategy (&TruncationStrategy): strategy to follow for the truncation, if required

stride (usize): amount of tokens to shift the input by if truncation is required (allowing for the generation of overlapping sequences with overflowing tokens)

Returns

Vec<TokenizedInput> containing the encoding output (token indices, token types, segment ids, ovrflowing tokens and special token mask) for each provided text

Example

fn clean_up_tokenization(&self, input_string: String) -> String

Cleans-up tokenization artifacts (for example whitespace before punctuation)

Arguments

input_string (String): input string to clean up

Returns

String: clean-up string

Example

fn build_input_with_special_tokens(
&self,
tokens_ids_with_offsets_1: TokenIdsWithOffsets,
tokens_ids_with_offsets_2: Option<TokenIdsWithOffsets>
) -> TokenIdsWithSpecialTokens

Build model inputs from a sequence or a pair of sequence for sequence classification tasks by concatenating and adding special tokens.

For example, a RoBERTa sequence has the following format:

single sequence: X

pair of sequences: A B

Parameters

tokens_ids_with_offsets_1 (TokenIdsWithOffsets): first sequence

tokens_ids_with_offsets_2 (TokenIdsWithOffsets): (optional) second sequence

Returns

TokenIdsWithSpecialTokens containing a concatenation of both sequences with added special tokens

Example