Module tokenizers::vocab

Vocabularies

This module contains the vocabularies leveraged by the tokenizer. These contain methods for deserialization of vocabulary files and access by the tokenizers, including:

dictionaries (mapping from token to token ids)

The following vocabularies have been implemented:

BERT Vocabulary implement the Vocab trait exposing a standard interface for integration with

Re-exports

pub use bert_vocab::BertVocab;

Modules

bert_vocab

Traits

Vocab

Base Vocab trait