Trait tokenizers::vocab::Vocab

pub trait Vocab {
    fn values(&self) -> &BTreeMap<String, i64>;
    fn indices(&self) -> &BTreeMap<i64, String>;
    fn special_values(&self) -> &BTreeMap<String, i64>;
    fn special_indices(&self) -> &BTreeMap<i64, String>;
    fn token_to_id(&self, token: &str) -> i64;
    fn id_to_token(&self, id: i64) -> &str;

    fn _token_to_id(
        &self, 
        token: &str, 
        values: &BTreeMap<String, i64>, 
        special_values: &BTreeMap<String, i64>, 
        unknown_value: &str
    ) -> i64 { ... }
    fn _register_as_special_value(
        token: &str, 
        values: &BTreeMap<String, i64>, 
        special_values: &mut BTreeMap<String, i64>
    ) -> Result<(), TokenError> { ... }
    fn _id_to_token<'a>(
        &self, 
        id: i64, 
        values: &'a BTreeMap<i64, String>, 
        special_indices: &'a BTreeMap<i64, String>, 
        unknown_value: &'a str
    ) -> &'a str { ... }
}

Base Vocab trait

Defines a common interface to the vocabularies for use in the tokenizers.

Required Methods

fn values(&self) -> &BTreeMap<String, i64>

Return the map of token strings to IDs

fn indices(&self) -> &BTreeMap<i64, String>

fn special_values(&self) -> &BTreeMap<String, i64>

fn special_indices(&self) -> &BTreeMap<i64, String>

Return the map of token IDs to strings for special values

fn token_to_id(&self, token: &str) -> i64

Converts a token to an id.

Parameters

token (&str): token to convert

Returns

i64: token index for the value provided. If not found in the indices, returns the unknown token index

fn id_to_token(&self, id: i64) -> &str

Returns

String: token value for the index provided. If not found in the indices, returns the unknown token value

Provided Methods

fn _token_to_id(
&self,
token: &str,
values: &BTreeMap<String, i64>,
special_values: &BTreeMap<String, i64>,
unknown_value: &str
) -> i64

Converts a token to an id, provided a BTreeMap of values, a BTreeMap of special values and the unknown value token string representation. This is not meant to be directly used, the method token_to_id offers a more convenient interface for most vocabularies, but needs to be implemented by the specific vocabulary.

Parameters

token (&str): token to convert

values (&BTreeMap<String, i64>): mapping from tokens to ids

special_values (&BTreeMap<String, i64>): mapping from special tokens to ids

unknown_value (&str): unknown token value

Returns

i64: index value for the provided token

fn _register_as_special_value(
token: &str,
values: &BTreeMap<String, i64>,
special_values: &mut BTreeMap<String, i64>
) -> Result<(), TokenError>

Register a token as a special value

Parameters

token (&str): token to register as a special value

values (&BTreeMap<String, i64>): mapping from tokens to ids. This should contain the token to add and will be used to read the id for registration in special_values

special_values (&BTreeMap<String, i64>): mapping from special tokens to ids

fn _id_to_token<'a>(
&self,
id: i64,
values: &'a BTreeMap<i64, String>,
special_indices: &'a BTreeMap<i64, String>,
unknown_value: &'a str
) -> &'a str

Converts an id to a token, provided a HashMap of values, a HashMap of special values and the unknown value token string representation. This is not meant to be directly used, the method id_to_token offers a more convenient interface for most vocabularies, but needs to be implemented by the specific vocabulary.

Parameters

id (&i64): token id to convert

indices (&HashMap<i64, String>): mapping from tokens to ids

special_indices (&HashMap<i64, String>): mapping from special tokens to ids

unknown_value (&str): unknown token value

Returns

String: token value for the index provided. If not found in the indices, returns the unknown token value